+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For a full diff see [`0.6.0...master`][0.6.0...master].
 - Added a missing return type declaration to `NodeInterface::getSize()` ([#150]), by [@localheinz]
 - Added parameter type declarations ([#151]), by [@localheinz]
 - Added property type declarations ([#152]), by [@localheinz]
+- Returned empty array from `Node::getNeigbors()` when node is root ([#153]), by [@localheinz]
 
 ## [`0.6.0`][0.6.0]
 
@@ -226,6 +227,7 @@ For a full diff see [`fcfd14e...v0.1.1`][fcfd14e...0.1.1].
 [#150]: https://github.com/nicmart/Tree/pull/150
 [#151]: https://github.com/nicmart/Tree/pull/151
 [#152]: https://github.com/nicmart/Tree/pull/152
+[#153]: https://github.com/nicmart/Tree/pull/153
 
 [@asalazar-pley]: https://github.com/asalazar-pley
 [@Djuki]: https://github.com/Djuki